+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog of lizard-graph
 0.3 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Added option for legend-location.
 
 
 0.2 (2011-12-08)